src/agent/requirements.txt
# Protocol OMNI v16.3.0 - Operation Eagle Eye (2026-01-24)
# Sentinel Audit: All versions verified against PyPI current releases

fastapi==0.128.0
uvicorn[standard]>=0.32.0
httpx>=0.28.0
pydantic>=2.10.0

# Phase 3: Deep Observability - OpenTelemetry + OpenInference
opentelemetry-api>=1.39.1
opentelemetry-sdk>=1.39.1
opentelemetry-exporter-otlp>=1.39.1
openinference-instrumentation-openai>=0.1.0
arize-phoenix-otel>=0.14.0

# Phase 4: Sovereign Cognition - LangGraph + Memory
langgraph>=1.0.3
langchain-core>=0.3.0
mem0ai>=1.0.2
